fix: hub v2 is not compatible with tensorflow v1
